requirements.txt
# pytorch==2.0.0
torchvision==0.15.0
torchaudio==2.0.0
omegaconf
torchmetrics==0.10.3
fvcore
iopath
xformers==0.0.16
huggingface-hub==0.20.2
h5py
numpy
pandas
pillow
tqdm
einops
webdataset
matplotlib
lifelines
scikit-survival
scikit-learn
tensorboard
fairscale
wandb
timm==1.0.14
packaging==23.2
ninja==1.11.1.1
transformers==4.36.2
monai
openslide-python
scikit-image